Storage module reorganization
Split up storage module to be easier
to understand.
